refactor: update interaction submodule names
* Update file names for Interaction subclasses -> suggestion by EmreTech
